move components into better folders
