+++ lib/platform-bible-react/src/components/advanced/extension-marketplace/buttons/install-button.component.tsx
