+++ lib/platform-bible-react/src/components/advanced/extension-marketplace/buttons/update-button.component.tsx
