+++ lib/platform-bible-react/src/components/advanced/extension-marketplace/footer.component.tsx
