+++ lib/platform-bible-react/src/components/advanced/extension-marketplace/markdown-renderer.component.tsx
