+++ lib/platform-bible-react/src/components/advanced/extension-marketplace/more-info.component.tsx
