+++ lib/platform-bible-react/src/components/advanced/scripture-results-viewer/scripture-results-viewer.component.css
