+++ lib/platform-bible-react/src/preview/pages/components/advanced.component.tsx
