+++ lib/platform-bible-react/src/preview/pages/components/advanced/inventory-example.component.tsx
