+++ lib/platform-bible-react/src/preview/pages/components/advanced/scripture-snippet.ts
