+++ lib/platform-bible-react/src/preview/pages/components/basics.component.tsx
