+++ lib/platform-bible-react/src/preview/pages/components/basics/checkbox.examples.component.tsx
