+++ lib/platform-bible-react/src/preview/pages/components/basics/slider.examples.component.tsx
