+++ lib/platform-bible-react/src/preview/pages/components/basics/switch.examples.component.tsx
