+++ lib/platform-bible-react/src/preview/pages/components/basics/tab.examples.component.tsx
