+++ lib/platform-bible-react/src/preview/pages/components/basics/table.examples.component.tsx
