+++ lib/platform-bible-react/src/preview/pages/guide.component.tsx
