+++ lib/platform-bible-react/src/preview/pages/guide/direction.guide.component.tsx
