+++ lib/platform-bible-react/src/preview/pages/guide/theme-color-display.component.tsx
