+++ lib/platform-bible-react/src/preview/pages/layouts.component.tsx
