+++ lib/platform-bible-react/src/preview/pages/layouts/dashboard5.layout.component.tsx
