+++ lib/platform-bible-react/src/preview/pages/layouts/window.layout.component.tsx
